Compile Window Counts

Query and compile window count results using the queryWindowCnts function.

getWindowCounts(
  dam = c("LWG", "WFF", "BON", "TDA", "JDA", "MCN", "IHR", "LMN", "LGS", "PRO", "ROZ",
    "PRD", "WAN", "RIS", "TUM", "RRH", "WEL", "ZOS"),
  spp = c("Chinook", "Coho", "Sockeye", "Steelhead", "Wild_Steelhead", "Shad",
    "Jack_Chinook", "Jack_Coho", "Jack_Sockeye", "Jack_Steelhead", "Lamprey",
    "Bull_Trout"),
  start_date = NULL,
  end_date = NULL,
  incl_jacks = F,
  sthd_type = c("all", "unclipped")
)

Arguments

dam

the dam code for the dam you wish to query for window counts. Possible codes are: WFF (Willamette Falls), BON (Bonneville), TDA (The Dalles), JDA (John Day), MCN (McNary), IHR (Ice Harbor), LMN (Lower Monumental), LGS (Little Goose), LWG (Lower Granite), PRO (Prosser), ROZ (Roza), PRD (Priest Rapids), WAN (Wanapum), RIS (Rock Island), TUM (Tumwater), RRH (Rocky Reach), WEL (Wells), ZOS (Zosel)

start_date

character vector of date (YYYYMMDD) when query should start

end_date

character vector of date (YYYYMMDD) when query should end

incl_jacks

should jacks be included in the window count totals? T / F

sthd_type

should window counts of steelhead be for all steelhead, or only unclipped (i.e. wild) fish? Default is all.

Author

Kevin See

Examples

getWindowCounts(start_date = '20150301', end_date = '20150817')
#> Rows: 174 Columns: 7
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (6): year, mm-dd, location, parameter, unit, datatype
#> dbl (1): value
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 170 × 4
#>    Species  Year Date       win_cnt
#>    <chr>   <int> <date>       <dbl>
#>  1 Chinook  2015 2015-03-01      NA
#>  2 Chinook  2015 2015-03-02      NA
#>  3 Chinook  2015 2015-03-03      NA
#>  4 Chinook  2015 2015-03-04      NA
#>  5 Chinook  2015 2015-03-05       0
#>  6 Chinook  2015 2015-03-06       0
#>  7 Chinook  2015 2015-03-07       0
#>  8 Chinook  2015 2015-03-08       0
#>  9 Chinook  2015 2015-03-09       0
#> 10 Chinook  2015 2015-03-10       0
#> # ℹ 160 more rows