PIT tag data

Query and download PIT tag data about night passage and re-ascension at particular dams for specific species and year, using DART.

queryPITtagData(
  damPIT = c("GRA", "PRA"),
  spp = c("Chinook", "Coho", "Steelhead", "Sockeye"),
  start_date = NULL,
  end_date = NULL
)

Arguments

damPIT
the dam code for the dam you wish to query for PIT tag data. Currently only available for Lower Granite Dam (GRA) and Priest Rapids (PRA).

spp
species to query PIT tag data for. Possible choices are: Chinook, Coho, Steelhead, Sockeye

start_date
character vector of date (YYYYMMDD) when query should start

end_date
character vector of date (YYYYMMDD) when query should end

Source

http://www.cbr.washington.edu/dart

Author

Kevin See

Examples

queryPITtagData(start_date = '20150701')

#> # A tibble: 53,370 x 25
#>    Ladder  Year Species SpCode TagId      WindowDetectionF… ClockHour Date      
#>    <chr>  <dbl> <chr>    <dbl> <chr>                  <dbl> <chr>     <date>    
#>  1 GRA     2016 Chinook      1 384.3B239…                 1 0900      2015-07-13
#>  2 GRA     2016 Chinook      1 384.3B239…                 0 0900      2015-07-13
#>  3 GRA     2016 Chinook      1 384.3B239…                 0 0900      2015-07-13
#>  4 GRA     2016 Chinook      1 384.3B239…                 0 0900      2015-07-13
#>  5 GRA     2016 Chinook      1 384.3B23A…                 1 0700      2015-07-14
#>  6 GRA     2016 Chinook      1 384.3B23A…                 0 0700      2015-07-14
#>  7 GRA     2016 Chinook      1 384.3B23A…                 0 0700      2015-07-14
#>  8 GRA     2016 Chinook      1 384.3B23A…                 0 0700      2015-07-14
#>  9 GRA     2016 Chinook      1 384.3B239…                 1 0900      2015-07-15
#> 10 GRA     2016 Chinook      1 384.3B239…                 0 0900      2015-07-15
#> # … with 53,360 more rows, and 17 more variables: WindowCountHours <chr>,
#> #   Period <chr>, RearType <chr>, TagStage <chr>, Coil <chr>,
#> #   DetectionDateTime <dttm>, TravelTime <dbl>, TagIdAscentCount <dbl>,
#> #   ReleaseYear <dbl>, ReleaseDateTime <dttm>, Run <dbl>, Length <dbl>,
#> #   CaptureMethod <chr>, FlagCodes <chr>, ReleaseSite <chr>, TagFile <chr>,
#> #   ReleaseRKM <chr>