PIT tag data

Query and download data about PIT tags observed crossing a particular site, using DART.

queryPITtagObs(
  site = c("GRA", "PRA"),
  spp = c("Chinook", "Coho", "Steelhead", "Sockeye"),
  start_date = NULL,
  end_date = NULL
)

Source

https://www.cbr.washington.edu/dart

Arguments

site

The site to query observations for. Query only set up for Lower Granite Dam Adult Fishway (GRA) and Priest Rapids fishway (PRA) at this time.

spp

species to query PIT tag observations for. Possible choices are: `Chinook`, `Coho`, `Steelhead`, `Sockeye`

start_date

character vector of date (YYYYMMDD) when query should start

end_date

character vector of date (YYYYMMDD) when query should end

Author

Kevin See

Examples

queryPITtagObs(spp = "Steelhead", start_date = "20180701")
#> # A tibble: 3,954 × 18
#>     Year TagFile        NFish TagCode SpRrt Length MarkSite ReleaseSite TotalRkm
#>    <int> <chr>          <dbl> <chr>   <chr>  <dbl> <chr>    <chr>          <dbl>
#>  1  2018 JKF18166.AFF       2 3DD.00… 32U      710 BONAFF   BONAFF           234
#>  2  2018 JSW-2018-183-…     2 3DD.00… 32W      530 LGRLDR   LGRLDR           695
#>  3  2018 JSW-2018-183-…     2 3DD.00… 32W      720 LGRLDR   LGRLDR           695
#>  4  2018 JSW-2018-184-…     6 3DD.00… 32W      650 LGRLDR   LGRLDR           695
#>  5  2018 JSW-2018-184-…     6 3DD.00… 32W      660 LGRLDR   LGRLDR           695
#>  6  2018 JSW-2018-184-…     6 3DD.00… 32W      530 LGRLDR   LGRLDR           695
#>  7  2018 JSW-2018-184-…     6 3DD.00… 32W      760 LGRLDR   LGRLDR           695
#>  8  2018 JSW-2018-184-…     6 3DD.00… 32W      700 LGRLDR   LGRLDR           695
#>  9  2018 JSW-2018-185-…     1 3DD.00… 32W      580 LGRLDR   LGRLDR           695
#> 10  2018 JSW-2018-184-…     6 3DD.00… 32W      730 LGRLDR   LGRLDR           695
#> # ℹ 3,944 more rows
#> # ℹ 9 more variables: ReleaseDate <date>, ObsSite <chr>, ObsTime <dttm>,
#> #   Stage <chr>, TravelDays <dbl>, NotFirst <chr>, MinTime <dttm>,
#> #   MaxTime <dttm>, LocationDays <dbl>