DART Window Counts

Query and download window counts at a dam by species and year using DART

queryWindowCnts(
  dam = c("LWG", "WFF", "BON", "TDA", "JDA", "MCN", "IHR", "LMN", "LGS", "PRO", "ROZ",
    "PRD", "WAN", "RIS", "TUM", "RRH", "WEL", "ZOS"),
  spp_code = c("fc", "fk", "fb", "fs", "fsw", "fa", "fcj", "fkj", "fl", "ft"),
  start_date = NULL,
  end_date = NULL
)

Source

https://www.cbr.washington.edu/dart

Arguments

dam

the dam code for the dam you wish to query for window counts. Possible codes are: WFF (Willamette Falls), BON (Bonneville), TDA (The Dalles), JDA (John Day), MCN (McNary), IHR (Ice Harbor), LMN (Lower Monumental), LGS (Little Goose), LWG (Lower Granite), PRO (Prosser), ROZ (Roza), PRD (Priest Rapids), WAN (Wanapum), RIS (Rock Island), TUM (Tumwater), RRH (Rocky Reach), WEL (Wells), ZOS (Zosel)

spp_code

species code(s) to query window counts for. Possible codes are: fc (Chinook), fk (Coho), fb (Sockeye), fs (Steelhead), fsw (Wild Steelhead), fa (Shad), fcj (Jack Chinook), fkj (Jack Coho), fbj (Jack Sockeye), fsj (Jack Steelhead), fl (Lamprey), ft (Bull Trout

start_date

character vector of date (YYYYMMDD) when query should start

end_date

character vector of date (YYYYMMDD) when query should end

Author

Kevin See

Examples

queryWindowCnts(dam = "LWG", spp_code = "fc", start_date = '20150301', end_date = '20150817')
#> # A tibble: 170 × 4
#>    Location  Year Date       Chinook
#>    <chr>    <int> <date>       <dbl>
#>  1 LWG       2015 2015-03-01      NA
#>  2 LWG       2015 2015-03-02      NA
#>  3 LWG       2015 2015-03-03      NA
#>  4 LWG       2015 2015-03-04      NA
#>  5 LWG       2015 2015-03-05       0
#>  6 LWG       2015 2015-03-06       0
#>  7 LWG       2015 2015-03-07       0
#>  8 LWG       2015 2015-03-08       0
#>  9 LWG       2015 2015-03-09       0
#> 10 LWG       2015 2015-03-10       0
#> # ℹ 160 more rows