Compile weekly data for Lower Granite dam

Combine window counts and PIT tag data on weekly time-steps

summariseLGRweekly(
  wind_data = NULL,
  pit_all = NULL,
  trap_df = NULL,
  incl_clip_sthd = FALSE,
  sthd_B_run = FALSE,
  incl_chnk_jacks = TRUE
)

Arguments

wind_data

data.frame containing window count data to summarise, queried using queryWindCnts, and potentially combined across calendar years

pit_all

data.frame containing data about individual PIT tags detected at Lower Granite Dam, produced using queryPITtagData

trap_df

data.frame as created by readLGRtrapDB.

incl_clip_sthd

should clipped steelhead in the trap be included in this summary? Should match with the window counts. Default is FALSE.

incl_chnk_jacks

should window counts of Chinook jacks be included with adult counts? Default is TRUE

Author

Kevin See