STADEM Data Summary

Query and summarise data for STADEM

compileGRAdata(
  yr,
  spp = c("Chinook", "Steelhead"),
  dam = c("LWG", "WFF", "BON", "TDA", "JDA", "MCN", "IHR", "LMN", "LGS", "PRO", "ROZ",
    "PRD", "WAN", "RIS", "TUM", "RRH", "WEL", "ZOS"),
  start_date = NULL,
  end_date = NULL,
  incl_jacks = NULL,
  sthd_type = c("all", "unclipped"),
  damPIT = c("GRA", "PRA"),
  strata_beg = "Mon",
  last_strata_min = 3,
  sthd_B_run = FALSE,
  trap_dbase = NULL,
  incl_trapRate = T,
  useDARTrate = F,
  trap_rate_cv = 0,
  trap_rate_dist = c("beta", "logit")
)

Arguments

yr

spawn year.

dam

the dam code for the dam you wish to query for window counts. Possible codes are: WFF (Willamette Falls), BON (Bonneville), TDA (The Dalles), JDA (John Day), MCN (McNary), IHR (Ice Harbor), LMN (Lower Monumental), LGS (Little Goose), LWG (Lower Granite), PRO (Prosser), ROZ (Roza), PRD (Priest Rapids), WAN (Wanapum), RIS (Rock Island), TUM (Tumwater), RRH (Rocky Reach), WEL (Wells), ZOS (Zosel)

start_date

character vector of date (YYYYMMDD) when query should start

end_date

character vector of date (YYYYMMDD) when query should end

incl_jacks

should jacks be included in the window count totals? T / F

sthd_type

should window counts of steelhead be for all steelhead, or only unclipped (i.e. wild) fish? Default is all.

damPIT

the dam code for the dam you wish to query for PIT tag data. Currently only available for Lower Granite Dam (GRA).

strata_beg

3 letter code for the day of the week each weekly strata should begin on. Default value is 'Mon'.

last_strata_min

minimum length (in days) for the final strata. Default value is 3.

sthd_B_run

should numbers of B run steelhead be reported? These are defined as wild steelhead greater than 780mm in length. Default is FALSE.

trap_dbase

data frame object containing the GRA trapping data with identical data types as in tblLGDMasterCombineExportJodyW; required fields include MasterID, LGDNumPIT, CollectionDate, SRR, LGDSpecies, LGDRear, LGDLifeStage, LGDMarkAD, LGDValid, LGDFlmm, PTAGISSxCGRAObse.

useDARTrate

should the DART query for the trap rate be used? Default is FALSE, which implies the trap rate is estimated by PIT tags.

trap_rate_cv

constant coefficient of variation (CV) that should be applied to estimates of trap rate queried by DART. Default value is 0.

trap_rate_dist

distributional form for trap rate prior. beta returns alpha and beta parameters for beta distribution. logit returns mean and standard deviation in logit space.

Author

Kevin See

Examples

compileGRAdata(2012)
#> Error in compileGRAdata(2012): !is.null(start_date) is not TRUE